# Shortcut postings for matchers when empty postings are selected