@pstibrany 2023-01-04
Nice, but why send this to mimir-prometheus instead of upstream? I don't see anything that wouldn't be accepted upstream, and more changes in mimir-prometheus creates merging headaches in the future.