@colega 2023-01-05
Upstreaming: https://github.com/prometheus/prometheus/pull/11813